rename navBar folder
